@@ -1,19 +1,24 @@
 {
-    "name": "rych/bencode",
+    "name": "yatts/bencode",
     "type": "library",
-    "description": "Bencode serializer for PHP 5.3+",
+    "description": "Bencode serializer for PHP 7.4+",
     "keywords": ["bencode", "serialize"],
-    "homepage": "https://github.com/rchouinard/bencode",
+    "homepage": "https://github.com/YetAnotherTorrentTrackerSoftware/bencode",
     "license": "MIT",
     "authors": [
         {
             "name": "Ryan Chouinard",
             "email": "rchouinard@gmail.com",
             "homepage": "http://ryanchouinard.com"
+        },
+        {
+            "name": "Fionera",
+            "email": "fionera@fionera.de",
+            "homepage": "http://fionera.de"
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.4"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.7"
