@@ -2,7 +2,7 @@
 /**
  * Rych Bencode
  *
- * Bencode serializer for PHP 5.3+.
+ * Bencode serializer for PHP 7.4.
  *
  * @package   Rych\Bencode
  * @copyright Copyright (c) 2014, Ryan Chouinard
@@ -44,10 +44,10 @@ class Encoder
      * @param  mixed   $data The value to encode.
      * @return string  Returns the bencode encoded string.
      */
-    public static function encode($data)
+    public static function encode($data): string
     {
         if (is_object($data)) {
-            if (method_exists($data, "toArray")) {
+            if (method_exists($data, 'toArray')) {
                 $data = $data->toArray();
             } else {
                 $data = (array) $data;
@@ -55,9 +55,7 @@ class Encoder
         }
 
         $encoder = new self($data);
-        $encoded = $encoder->doEncode();
-
-        return $encoded;
+        return $encoder->doEncode();
     }
 
     /**
@@ -66,20 +64,24 @@ class Encoder
      * @param  mixed   $data The value to encode.
      * @return string  Returns the bencode encoded string.
      */
-    private function doEncode($data = null)
+    private function doEncode($data = null): string
     {
-        $data = is_null($data) ? $this->data : $data;
+        $data = $data ?? $this->data;
 
         if (is_array($data) && (isset ($data[0]) || empty ($data))) {
             return $this->encodeList($data);
-        } elseif (is_array($data)) {
-            return $this->encodeDict($data);
-        } elseif (is_integer($data) || is_float($data)) {
-            $data = sprintf("%.0f", round($data, 0));
-            return $this->encodeInteger($data);
-        } else {
-            return $this->encodeString($data);
         }
+
+        if (is_array($data)) {
+            return $this->encodeDict($data);
+        }
+
+        if (is_int($data) || is_float($data)) {
+            $data = sprintf('%.0f', round($data, 0));
+            return $this->encodeInteger($data);
+        }
+
+        return $this->encodeString($data);
     }
 
     /**
@@ -88,10 +90,10 @@ class Encoder
      * @param  integer $data The integer to be encoded.
      * @return string  Returns the bencode encoded integer.
      */
-    private function encodeInteger($data = null)
+    private function encodeInteger($data = null): string
     {
-        $data = is_null($data) ? $this->data : $data;
-        return sprintf("i%.0fe", $data);
+        $data = $data ?? $this->data;
+        return sprintf('i%.0fe', $data);
     }
 
     /**
@@ -100,10 +102,10 @@ class Encoder
      * @param  string  $data The string to be encoded.
      * @return string  Returns the bencode encoded string.
      */
-    private function encodeString($data = null)
+    private function encodeString($data = null): string
     {
-        $data = is_null($data) ? $this->data : $data;
-        return sprintf("%d:%s", strlen($data), $data);
+        $data = $data ?? $this->data;
+        return sprintf('%d:%s', strlen($data), $data);
     }
 
     /**
@@ -112,11 +114,11 @@ class Encoder
      * @param  array   $data The list to be encoded.
      * @return string  Returns the bencode encoded list.
      */
-    private function encodeList(array $data = null)
+    private function encodeList(array $data = null): string
     {
-        $data = is_null($data) ? $this->data : $data;
+        $data = $data ?? $this->data;
 
-        $list = "";
+        $list = '';
         foreach ($data as $value) {
             $list .= $this->doEncode($value);
         }
@@ -130,12 +132,12 @@ class Encoder
      * @param  array   $data The dictionary to be encoded.
      * @return string  Returns the bencode encoded dictionary.
      */
-    private function encodeDict(array $data = null)
+    private function encodeDict(array $data = null): string
     {
-        $data = is_null($data) ? $this->data : $data;
+        $data = $data ?? $this->data;
         ksort($data); // bencode spec requires dicts to be sorted alphabetically
 
-        $dict = "";
+        $dict = '';
         foreach ($data as $key => $value) {
             $dict .= $this->encodeString($key) . $this->doEncode($value);
         }
